Separate out MappingId
This changeset separates cuda-specific typed mapping ids on the path towards supporting multiple backends.
